% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{add.alpha}
\alias{add.alpha}
\title{add.alpha}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
add.alpha()
}
\keyword{cats}
