% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{add_rects}
\alias{add_rects}
\title{add_rects}
\usage{
add_rects(threshyellow, threshred, transp)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
add_rects()
}
\keyword{cats}
