% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{asplit}
\alias{asplit}
\title{asplit}
\usage{
asplit(data, var)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Splits a data into lists based on var input as split variable
}
\examples{
asplit()
}
\keyword{cats}
