% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bin_vector}
\alias{bin_vector}
\title{bin a numeric vector and return a factor of bins}
\usage{
bin_vector(numeric_vector, breaks = 10, binLabels = NULL, return_df = TRUE)
}
\arguments{
\item{numeric_vector}{}

\item{breaks}{either a single number specifying the number of equally widthed breaks, or a vector of break boundaries}

\item{binLabels}{optional labels for the new bins that are returned instead of the bin range}

\item{return_df}{return a data.frame with boundaries, the value and the bin level}
}
\description{
bin a numeric vector and return a factor of bins
}
\examples{
vec = sample(1:100, size = 100)
bin_vector(vec, breaks = 10)
bin_vector(vec, breaks = c(0, 33, 66, 100))
bin_vector(vec, breaks = c(0, 33, 66, 100), 
           binLabels = c('lower tertile', 'middle tertile', 'upper tertile'),
           return_df = FALSE)
}
