% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{calc_dewpoint}
\alias{calc_dewpoint}
\title{calc_dewpoint}
\usage{
calc_dewpoint(df, Tair_f_col, relhumCOL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
calc_dewpoint()
}
\keyword{cats}
