% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{completeOnCol}
\alias{completeOnCol}
\title{completeOnCol}
\usage{
completeOnCol(data, desiredCols)
}
\description{
This function takes a factor and coverts it to numeric
}
\examples{
completeOnCol()
}
\keyword{cats}
