% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{detachAllPackages}
\usage{
detachAllPackages()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
detachAllPackages()
}
\keyword{cats}
