% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{df_t}
\alias{df_t}
\title{df_t}
\usage{
df_t(x, Y)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
df_t()
}
\keyword{cats}
