% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{duplicates}
\alias{duplicates}
\title{duplicates}
\usage{
duplicates(data, oncols)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
duplicates()
}
\keyword{cats}
