% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{factorCOL}
\alias{factorCOL}
\title{factorCOL}
\usage{
factorCOL(dataTOfactor, newLEVELS = NULL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This creates an input factor and remaps levels to integer sequence
}
\examples{
factorCOL()
}
\keyword{cats}
