% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{from_to}
\alias{from_to}
\title{remap values}
\usage{
from_to(vec, old_new_map_vec)
}
\arguments{
\item{vec}{Vector of genomic region strings}

\item{old_new_map_vec}{named vector with names and values corresponding to old + new pairs of values}

\item{...}{Additional arguments passed to}
}
\description{
pass a named vector to replace values in a second vector
}
\examples{
#map values from old to new
from_to(vec = c('T_cell', 'B_cells_2', 'B_cell_1'),  
old_new_map_vec = c('B_cells_1' = 'B cell','B_cells_2' = 'B cell'))
}
