% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{get3days}
\alias{get3days}
\title{get3days}
\usage{
get3days(df, whichCOLS, splitVAR)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
get3days()
}
\keyword{cats}
