% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_mode}
\alias{get_mode}
\title{get the mode}
\usage{
get_mode(vec)
}
\arguments{
\item{vec}{}
}
\description{
On equal numbers of occurrence return first item
}
\examples{
get_mode(c('A','B','D','D','A'))
}
