% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{getmeans}
\alias{getmeans}
\title{getmeans}
\usage{
getmeans(data)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
getmeans()
}
\keyword{cats}
