% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{group_boxplots}
\alias{group_boxplots}
\title{group_boxplots}
\usage{
group_boxplots(dataog, VAR, groupVAR, horiz = TRUE, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
group_boxplots()
}
\keyword{cats}
