% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{heat.index2}
\alias{heat.index2}
\title{heat.index2}
\usage{
heat.index2(t = NA, dp = c(), rh = c(),
  temperature.metric = "fahrenheit", output.metric = NULL, round = 0)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
heat.index2()
}
\keyword{cats}
