% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{heat_index_2}
\alias{heat_index_2}
\title{heat_index_2}
\usage{
heat_index_2(t = NA, rh = NA)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
heat_index_2()
}
\keyword{cats}
