% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{ipak}
\alias{ipak}
\title{ipak}
\usage{
ipak(pkg)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
ipak()
}
\keyword{cats}
