% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j.R
\name{j}
\alias{j}
\title{print small subset of R object as overview}
\usage{
j(obj, headtail = NULL, print_nested = FALSE, show_max = 5)
}
\arguments{
\item{obj}{Object, for which an overview is desired}

\item{headtail}{return specified number of items from the start and the end of the object (not lists). Can result in duplicated items when headtail is bigger than the object length.}

\item{print_nested}{also print overview for lists within lists. Default=FALSE, Beta feature!}
}
\description{
Works for all R objects in contrast to head().
}
\examples{

}
\keyword{glimpse}
\keyword{head,}
\keyword{overview,}
