% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{jj}
\alias{jj}
\title{print small subset of R object as overview}
\usage{
jj(obj, headtail = NULL)
}
\arguments{
\item{obj}{Object, for which an overview is desired}
}
\description{
Works for all R objects in contrast to head().
}
\examples{

}
\keyword{glimpse}
\keyword{head,}
\keyword{overview,}
