% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_arrange_ggplots.R
\name{jj_arrange_ggplots}
\alias{jj_arrange_ggplots}
\title{arrange ggplots in a grid}
\usage{
jj_arrange_ggplots(
  gg_list,
  nplots = 4,
  cols = 2,
  legend.pos = NULL,
  equal_size_plots = TRUE
)
}
\arguments{
\item{gg_list}{list of ggplot objects}

\item{nplots}{number of plots to combine together}

\item{cols}{number of columns for the plot layout}

\item{legend.pos}{position of legends. One of 'left','right','top','bottom','none'. If null, do not change legend positions}

\item{equal_size_plots}{if TRUE, make all individual plots the same size}
}
\description{
plot multiple plots together in a grid based on a list of ggplot objects
The following error can occur and is related to problems with the legend. Try to change legend.pos to solve it:
Error in grid.Call(C_convert, x, as.integer(whatfrom), as.integer(whatto),  : Viewport has zero dimension(s)
}
\examples{
df = data.frame(a=rnorm(100, 0, 5), b=rnorm(100, 0, 5), d=rbinom(100, 50, 0.3), e = sample(c('A','B', 'C'), 100, replace=T))
ggl = list()
ggl[[1]] = ggplot(df, aes(x = a, y = b, colour=d)) + geom_point()
ggl[[2]] = ggplot(df, aes(x = a, y = b, colour=e)) + geom_point()
ggl[[3]] = ggplot(df, aes(x = a)) + geom_histogram()
jj_arrange_ggplots(ggl, nplots = 4, cols = 2)
jj_arrange_ggplots(ggl, nplots = 4, cols = 2, legend.pos = 'none')
}
\keyword{ggplot,plot_grobs,plot_grid}
