% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_plot_association_heatmap.R
\name{jj_association_heatmap}
\alias{jj_association_heatmap}
\alias{jj_metadata_association_heatmap}
\title{find associations of embedding scores with meta data}
\usage{
jj_metadata_association_heatmap(
  score_mat,
  metadat_mat,
  plotAssocFor = NULL,
  cor_cutoff = 0.2,
  categorical_skip = 30,
  ...
)
}
\arguments{
\item{metadat_mat}{data.frame/matrix with the metadata for each cell}

\item{plotAssocFor}{Option to specify subset of metadata_mat for which associations should be calulated}

\item{cor_cutoff}{Only show correlations with absolute value above this cutoff for better overview}

\item{categorical_skip}{Skip categorical metadata columns with more than this number of distinct values}

\item{scores_mat}{numeric matrix of scores (eg pc scores)}

\item{group_vec}{vector of length = nrow(scores_mat) with group annotation}
}
\description{
jj_metadata_association_heatmap: Plot heatmap to find metadata associations with a set of scores, e.g. from PCA
jj_association_heatmap: Plot mean or scaled mean per group for all columns in scores_mat, good to further understand single metadata associations with scores.
Further arguments can be passed to ComplexHeatmap
}
\examples{
#e.g. PC2 has positive correlation with the grouping from RNA_snn_res.1
pca_scores = Embeddings(pbmc_small, 'pca')[,1:10]
jj_metadata_association_heatmap(score_mat = pca_scores, metadat_mat = pbmc_small[[]],
                                cor_cutoff = 0.2)
# Groups 1 and 2 tend to have high scores and group 0 has low scores for PC2
jj_association_heatmap(scores_mat = pca_scores, group_vec = pbmc_small$RNA_snn_res.1)
# confirm by plotting the data
jj_plot_features(pbmc_small, reduction = 'pca', meta_features = 'RNA_snn_res.1', pt.size = 2)
}
