% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_bind_features_with_dr_df.R
\name{jj_bind_features_with_dr_df}
\alias{jj_bind_features_with_dr_df}
\title{bind features together with meta.data}
\usage{
jj_bind_features_with_dr_df(
  obj,
  assay = "RNA",
  slot = "counts",
  features,
  dr_df = NULL,
  cap_top = NULL,
  cap_bottom = NULL,
  log10Transform = FALSE,
  ...
)
}
\arguments{
\item{obj}{Seurat object or matrix of features*samples}

\item{assay}{assay in Seurat object to use}

\item{slot}{slot in assay to use}

\item{features}{Features to extract from obj}

\item{dr_df}{optional data.frame, which is joined together with the feature matrix}

\item{cap_top}{value passed to jj_cap_vals}

\item{cap_bottom}{value passed to jj_cap_vals}

\item{log10Transform}{if TRUE, return log10 transformed feature values, default: F}
}
\description{
add values from a matrix/seurat object to data.frame containing metadata (dr_df)
}
\examples{
jj_bind_features_with_dr_df(seurat_rna, assay='RNA', slot='data', cap_top='q95')
}
\keyword{cap}
