% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_cap_vals.R
\name{jj_cap_vals}
\alias{jj_cap_vals}
\title{Cap values at the top/bottom of a vector by a threshold}
\usage{
jj_cap_vals(vec, cap_top = "q99", cap_bottom = NULL)
}
\arguments{
\item{vec}{numeric vector}

\item{cap_top}{cap values above this threshold, either a single number or a quantile as string, eg. 'q95' for quantile 95}

\item{cap_bottom}{same as cap_top, but cap values below this threshold}
}
\description{
Useful for plotting when there are few outlier values stretching the colour scale
}
\examples{
jj_cap_vals(seq(0,100), cap_top='q95', cap_bottom=10)
}
\keyword{cap}
