% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_combine_plots.R
\name{jj_combine_plots}
\alias{jj_combine_plots}
\title{arrange ggplots into combined plots of n plots}
\usage{
jj_combine_plots(gg_list, nplots = 4, cols = 2)
}
\arguments{
\item{gg_list}{list of ggplot objects}

\item{nplots}{number of ggplots to plot together in one combined plot}

\item{cols}{number of columns for each combined plot}
}
\description{
arrange ggplots into combined plots of \code{n} plots and \code{cols} columns
}
\examples{
gg_list = lapply(1:6, function(x) ggplot(data = data.frame(a=rnorm(10, 0, 5), b=rnorm(10, 0, 5)), aes(x=a,y=b))+geom_point(colour=x))
jj_combine_plots(gg_list, nplots=6, cols = 4)
}
\keyword{plot}
