% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_compare_vectors.R
\name{jj_compare_vectors}
\alias{jj_compare_vectors}
\title{quantify number of common elements between two or more vectors}
\usage{
jj_compare_vectors(A, B)
}
\arguments{
\item{A}{first vector}

\item{B}{second vector}

\item{list_of_vectors}{list of vectors to be compared}

\item{mainbar.y.label}{Passed to upset function, default: 'Intersection Size',}

\item{mb.ratio}{Passed to upset function, default: c(0.55, 0.45)}
}
\description{
jj_compare_vectors returns a data.frame specifying how many elements of vector A are present in B and vice versa
jj_plot_upsetr uses the UpSetR:upset function to plot overlaps between vectors passed as a list
jj_make_overlap_df returns a list: key_table mapping the names of the list to letters and overlaps, a data.frame with the elements that are present in each combination of vectors passed as a list
}
\examples{
vec_list = list(group1= 1:5, group2= 2:4, group3 = c(3,6), group4=5:10, group5=12)
jj_compare_vectors(A=vec_list[[1]], B=vec_list[[2]])
jj_plot_upsetr(vec_list)
jj_make_overlap_df(vec_list)
}
