% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_describe_df.R
\name{jj_describe_df}
\alias{jj_describe_df}
\title{describe all the columns in a data.frame}
\usage{
jj_describe_df(df, convert = FALSE, return_df = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{convert}{if convert = T, convert character columns that only contain digits to numeric columns before summarising}

\item{return_df}{if return_df = T, return the summary tibble instead of printing it out}
}
\description{
Gives an overview of a data.frame (dimensions, class of each column, mode/min/mean/max, distinct values, na_count, table, head)
}
\examples{
jj_describe_df(pbmc_small[[]])
}
