% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_filter_values_by_abundance.R
\name{jj_filter_values_by_abundance}
\alias{jj_filter_values_by_abundance}
\title{fitler a list of vectors for those values having min/max abundance}
\usage{
jj_filter_values_by_abundance(
  list_of_vectors,
  max_abundance = NULL,
  min_abundance = NULL
)
}
\arguments{
\item{list_of_vectors}{list of vectors, which are compared against each other}

\item{max_abundance}{only return elements that occur maximum 'max_abundance' times in the list (only counting once per vector)}

\item{min_abundance}{only return elements that occur minimum 'min_abundance' times in the list (only counting once per vector)

example_list = list(l1 = letters\link{1:7}, l2 = letters\link{1:4}, l3 = c( 'd', 'd', letters\link{4:6}))
jj_filter_values_by_abundance(example_list, min_abundance = 3)
jj_filter_values_by_abundance(example_list, max_abundance = 2)
jj_filter_values_by_abundance(example_list, min_abundance = 2, max_abundance = 3)}
}
\description{
fitler a list of vectors for those values having min/max abundance
}
