% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_get_colours.R
\name{jj_get_colours}
\alias{jj_get_colours}
\title{get colours as named vector from file}
\usage{
jj_get_colours(annotation_vector = NULL, colour_csv, plot_colours = FALSE)
}
\arguments{
\item{annotation_vector}{vector with the groups for which colours are needed}

\item{colour_csv}{two column csv with annotation in first column and colour in second column}

\item{plot_colours}{visualize colours as ggplot heatmap, default: F}
}
\description{
get colours as named vector from file
}
\examples{
#Save this in file called `colour_map.csv`
#/fruits
#banana,yellow
#apple,green
#cherry,red
#coconut,#D7B5D8
jj_get_colours(colour_csv='colour_map.csv', plot_colours=TRUE)
DimPlot(seurat_rna,  group.by = 'fruits') + scale_colour_manual(values=jj_get_colours(seurat_rna$clinical, colour_csv='colour_map.csv'))
}
\keyword{colours}
