% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_get_feature_mat.R
\name{jj_get_feature_mat}
\alias{jj_get_feature_mat}
\title{extract expression values for features of interest from feature matrix}
\usage{
jj_get_feature_mat(feature_mat, features = NULL, verbose = TRUE, ...)
}
\arguments{
\item{feature_mat}{matrix with n features (rows) and m observations (columns)}

\item{features}{features for which the values should be extracted from the matrix}
}
\description{
extracts gene expression values for a given vector of features of interest
}
\examples{

}
\keyword{expression,}
\keyword{extract}
\keyword{features}
\keyword{get}
