% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_get_feature_mat.R
\name{jj_get_features}
\alias{jj_get_features}
\title{extract expression values for features of interest from feature matrix}
\usage{
jj_get_features(
  feature_mat,
  features,
  fail_on_missing = FALSE,
  use_features_order = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{feature_mat}{matrix with n features (rows) and m observations (columns)}

\item{features}{features for which the values should be extracted from the matrix}

\item{fail_on_missing}{return error if not all features are found}

\item{use_features_order}{order the rows in the matrix by the order of \code{features}, otherwise just leave the order as in the supplied \code{feature_mat}}
}
\description{
extracts gene expression values for a given vector of features of interest
}
\examples{
mat_use = matrix(1:20, nrow=10, byrow=T)
rownames(mat_use) = LETTERS[1:10]
jj_get_features(mat_use, c('D','J','A'))
jj_get_features(mat_use, c('D','J','A', 'Z'),  use_features_order=F)
}
\keyword{expression,}
\keyword{extract}
\keyword{features}
\keyword{get}
