% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_get_reduction_coords.R
\name{jj_get_reduction_coords}
\alias{jj_get_reduction_coords}
\title{get cell embeddings together with meta data}
\usage{
jj_get_reduction_coords(obj, redname = NULL, exact_match = TRUE)
}
\arguments{
\item{obj}{Seurat, cisTopic or ArchR object}

\item{redname}{name of the reduction, for which cell embeddings should be obtained}

\item{exact_match}{if TRUE, \code{redname} must exactly match the name of the reduction in the object}
}
\description{
Extract cell embeddings together with meta data from a Seurat, cisTopic or ArchR-object
}
\examples{
head(jj_get_reduction_coords(pbmc_small, 'tsne'))
}
\keyword{embeddings}
