% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_gg_heatmap.R
\name{jj_gg_heatmap}
\alias{jj_gg_heatmap}
\title{feature heatmap based on ggplot}
\usage{
jj_gg_heatmap(
  df,
  x_group,
  y_group,
  value,
  y_annot = NULL,
  x_annot = NULL,
  order_by_annot = FALSE,
  cluster_columns = TRUE,
  cluster_rows = T,
  show_numbers = FALSE,
  size_x = 8,
  size_y = 10,
  col_scale = "rwb0"
)
}
\arguments{
\item{df}{data.frame in long format}

\item{x_group}{Group plotted in rows}

\item{y_group}{Group plotted in columns}

\item{value}{Continuous feature used to colour the tiles}

\item{y_annot}{Optional group used to annotate the columns (be careful not to mix up with rows)}

\item{x_annot}{Optional group used to annotate the rows (be careful not to mix up with columns)}

\item{order_by_annot}{instead of clustering columns by values, cluster columns by mixedorder of y annotation}

\item{cluster_columns}{logical, should rows be clustered}

\item{cluster_rows}{logical, should rows be clustered}

\item{show_numbers}{Plot the values as number in each tile}

\item{size_x}{size of x axis labels}

\item{size_y}{size of y axis labels}

\item{col_scale}{Colorscale to use, one of rwb0 (red <0, white = 0, blue >0), viridis}
}
\description{
feature heatmap based on ggplot
}
\examples{
data('mtcars')
mtcars2 = mtcars \%>\% dplyr::group_by(cyl, gear) \%>\% dplyr::summarise(mpg = mean(mpg))
mtcars2$cyl_annot = from_to(mtcars2$cyl, c('4' = 'few', '6' = 'many', '8' = 'many'))
jj_gg_heatmap(mtcars2, x_group = 'cyl', y_group = 'gear', value='mpg', show_numbers = T, col_scale = 'viridis', x_annot = 'cyl_annot')
}
