% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_initialize_df.R
\name{jj_initialize_df}
\alias{jj_initialize_df}
\title{create data.frame of defined size with initial value in each cell}
\usage{
jj_initialize_df(ncol, nrow, init = NA, col.names = NULL, ...)
}
\arguments{
\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{init}{value to put into every cell}

\item{col.names}{colnames to set}

\item{row.names}{rownames to set}
}
\description{
Creates a data.frame of size \code{ncol} * \code{nrow} filled with \code{init}. \code{row.names} and \code{column.names} may also be specified.
}
\examples{
jj_initialize_df(ncol = 3, nrow = 4, init = 0, col.names = paste0("col", 1:3), row.names = paste0("row", 1:4))
}
\keyword{create}
\keyword{data.frame}
\keyword{data.frame,}
\keyword{dummy}
\keyword{initialize}
