% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_plot_blended_feature.R
\name{jj_plot_blended_feature}
\alias{jj_plot_blended_feature}
\title{Plot scatterplot of two features as combined colorscale}
\usage{
jj_plot_blended_feature(df, x, y, colfeat1, colfeat2, pt_size = 1)
}
\arguments{
\item{df}{sparse matrix, vector, or data.frame}

\item{x}{vector with group annotation, has to have length equal to ncol(sparse_mat)/nrow(data.frame)/length(vector)}

\item{y}{provide \code{symbol_column} to add text to all points that are highlighted (when use_text=TRUE)}

\item{colfeat1}{vector specifying range of x axis (and for fc_fc plot range of y axis)}

\item{colfeat2}{define \code{marker_thres} to highlight all points with absolute value above}

\item{pt_size}{size of points}
}
\value{
ggplot scatterplot
}
\description{
jj_plot_blended_feature translates two continuous feature scales to a red and green colorscale and plots a blended colorscale a scatterplot.
Useful e.g. to visualize co-expression patterns.
}
\examples{
library(Seurat)
dr_df = jj_get_reduction_coords(pbmc_small, 'tsne')
dr_df = jj_bind_features_with_dr_df(pbmc_small, assay= 'RNA', slot = 'data', features = c('CD19', 'CD79A'), dr_df = dr_df, cap_top = 'q95')
jj_plot_blended_feature(dr_df, x = 'tSNE_1', y = 'tSNE_2', colfeat1 = 'CD19', colfeat2 = 'CD79A', pt_size = 3)
}
