% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_confusion_matrix.R
\name{jj_plot_confusion_matrix}
\alias{jj_plot_confusion_matrix}
\title{confusion_matrix}
\usage{
jj_plot_confusion_matrix(
  vector_a,
  vector_b,
  xlab_text = "vector_a",
  ylab_text = "vector_b",
  plot_absolute = F,
  plot_numbers = F,
  text_size = 2
)
}
\arguments{
\item{vector_a}{first vector, e.g. true label}

\item{vector_b}{second vector, e.g. predicted label}

\item{xlab_text}{x label}

\item{ylab_text}{y label}

\item{plot_absolute}{if FALSE, show the relative fraction of values for each cell}

\item{plot_numbers}{show count in fields}

\item{text_size}{size of text}
}
\value{
Returns a confusion matrix as heatmap
}
\description{
plot a confusion matrix based on two vectors of equal length
}
\examples{
true_label = c('A','A','A','B','B','B')
predicted_label = c('A','B','B','B','B','B')
#percentage of the total observations that are in the respective field
jj_plot_confusion_matrix(vector_a = true_label, vector_b = predicted_label,
                        xlab_text = 'true value', ylab_text = 'predicted value', 
                        plot_numbers = T, text_size = 10)
# absolute
jj_plot_confusion_matrix(vector_a = true_label, vector_b = predicted_label,
                        xlab_text = 'true value', ylab_text = 'predicted value', 
                        plot_numbers = T, text_size = 10, plot_absolute = T)
}
