% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_plot_n.R
\name{jj_plot_n}
\alias{jj_plot_n}
\title{arrange ggplots into combined plots of n plots}
\usage{
jj_plot_n(gg_list, n = 4, cols = 2)
}
\arguments{
\item{gg_list}{list of ggplot objects}

\item{n}{number of ggplots to plot together in one combined plot}

\item{cols}{number of columns for each combined plot}
}
\description{
arrange ggplots into combined plots of \code{n} plots and \code{cols} columns
}
\examples{
gg_list = lapply(1:10, function(x) ggplot(data = data.frame(a=rnorm(10, 0, 5), b=rnorm(10, 0, 5)), aes(x=a,y=b))+geom_point(colour=x))
jj_plot_n(gg_list,  n=10, cols = 4)
}
\keyword{plot}
