% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_prime_factorization.R
\name{jj_prime_factorization}
\alias{jj_prime_factorization}
\title{Prime factorization}
\usage{
jj_prime_factorization(x)
}
\arguments{
\item{x}{single number, which is to be split into prime numbers}
}
\description{
Returns the prime numbers as vector
}
\examples{
jj_prime_factorization(45)
}
\keyword{prime}
