% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_load_save_bed.R
\name{jj_save_bed}
\alias{jj_save_bed}
\title{write Granges or region string vector to bed file}
\usage{
jj_save_bed(regions, file_name)
}
\arguments{
\item{regions}{regions to save to bed file, either as GRanges or as strings in the format 'chr1-1-99'}

\item{file_name}{file name to save to or load from}

\item{return_granges}{if TRUE, return a GRanges object instead of a tibble when loading data}
}
\description{
estimate densities to include density area in tsne/umap
}
\examples{
jj_save_bed(c('chr1-234-3488', 'chrX-300-1500'), 'test.bed')
jj_load_bed('test.bed', return_granges = TRUE)
}
