% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_load_save_excel.R
\name{load_save_excel}
\alias{load_save_excel}
\alias{jj_load_excel}
\alias{jj_save_excel}
\title{Load/save excel workbooks using openxlsx library}
\usage{
jj_load_excel(file_name, sheet_names = NULL, row_names = FALSE, ...)

jj_save_excel(
  list_of_df,
  file_name,
  sheet_names = NULL,
  row_names = FALSE,
  ...
)
}
\arguments{
\item{file_name}{Full path to the file to read or write}

\item{sheet_names}{Names of sheets to read or write, default: all sheets when loading, names(list) when saving}

\item{row_names}{Set first column as rownames or write rownames from each df, default: FALSE}

\item{...}{further arguments passed to read.xlsx or write.xlsx}

\item{list_of_df}{list of data.frames to write into excel workbook}
}
\description{
load excel workbook as list of data.frames or save list of data.frames as excel
}
\keyword{excel}
