% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_to_vector}
\alias{matrix_to_vector}
\title{convert a matrix to a 1d vector}
\usage{
matrix_to_vector(mat, by = "column")
}
\arguments{
\item{mat}{matrix}

\item{by}{one of \code{column}, \code{row}: concatenate column- or rowwise}
}
\description{
convert a matrix to a 1d vector
}
\examples{
test_mat = matrix(1:12, ncol=4, byrow = FALSE)
test_mat
matrix_to_vector(test_mat)
matrix_to_vector(test_mat, by ='row')
}
