% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{mav}
\alias{mav}
\title{mav}
\usage{
mav(x, n)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
mav()
}
\keyword{cats}
