% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{max_solar}
\alias{max_solar}
\title{solar_bras}
\usage{
max_solar(df, lat, lon)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
add.alpha()
}
\keyword{cats}
