% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mixsort_factor}
\alias{mixsort_factor}
\title{alphanumerically sort levels of a factor}
\usage{
mixsort_factor(vec)
}
\arguments{
\item{vec}{vector to sort}
}
\description{
Return a factor with alphabetically and numerically sorted levels
}
\examples{
mixsort_factor(c('Sample_3', 'Sample_1', 'Sample_3', 'Sample_2'))
}
