% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{plotCI}
\alias{plotCI}
\title{plotCI}
\usage{
plotCI(x, y = NULL, uiw, liw = uiw, ..., sfrac = 0.01, vertLineLWD,
  confBoundsLWD)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
plotCI()
}
\keyword{cats}
