% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{solar_bras}
\alias{solar_bras}
\title{add.alpha}
\usage{
solar_bras(df, lat, lon)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
solar_bras()
}
\keyword{cats}
