% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{stationpressure}
\alias{stationpressure}
\title{stationpressure}
\usage{
stationpressure(df, presCOL, tairFcol, elevation_ft)
}
\description{
This function takes a factor and coverts it to numeric
}
\examples{
stationpressure()
}
\keyword{cats}
