% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{subsetNA}
\alias{subsetNA}
\title{subsetNA}
\usage{
subsetNA(data)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
subsetNA()
}
\keyword{cats}
