% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{subsetNA_DT}
\alias{subsetNA_DT}
\title{subsetNA_DT}
\usage{
subsetNA_DT(data, col)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
subsetNA_DT()
}
\keyword{cats}
