% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_histogram.R, R/jj_summarize_obj.R,
%   R/jj_volcano.R
\name{summarise_vals}
\alias{summarise_vals}
\alias{jj_plot_hist_wo0}
\alias{jj_summarize_sparse_mat}
\alias{jj_summarize_vector}
\alias{jj_summarize_dataframe}
\alias{jj_volcano_plot}
\title{summarise values as mean, mode, or sum by a group vector}
\usage{
jj_plot_hist_wo0(
  feature_vec,
  thres = NULL,
  title = "",
  xlab = "value",
  cols = c("red", "lightblue"),
  plot_numbers = 1:3,
  pad = 20
)

jj_summarize_sparse_mat(
  summarize_obj,
  summarize_by_vec,
  method = "mean",
  return_matrix = TRUE
)

jj_summarize_vector(
  summarize_obj,
  summarize_by_vec,
  method = "mean",
  order = T,
  return_vec = F
)

jj_summarize_dataframe(summarize_obj, summarize_by_vec, method = "mean")

jj_volcano_plot(
  plot_df,
  logfc_column,
  pval_column,
  symbol_column = NULL,
  marker_thres = c(0.5, Inf),
  labs_range = c(-1, 1),
  alpha = 1,
  pt.size = 2.5,
  highlight.pt.size = 2.5,
  col_vec = c("red", "blue", "black", "green"),
  markers_highlight = NULL,
  markers_highlight_col = NULL,
  only_highlight = FALSE,
  col_by_highlight = FALSE,
  use_text = TRUE,
  add_thres_line = FALSE,
  group_names = NULL,
  arrow_pos = list(x = 0.25, xend = 1, y = 40)
)
}
\arguments{
\item{feature_vec}{vector with values to plot}

\item{thres}{optional value threshold, for which the number of observations are counted above and below}

\item{title}{optional title}

\item{xlab}{xlab text}

\item{cols}{colours for bars above and below threshold}

\item{plot_numbers}{NULL (do not plot numbers) or vector containing any combination of 1 (observations above threshold), 2 (observations below threshold) and 3 (observations below threshold that are not 0)}

\item{pad}{padding for the numbers if plot_numbers is not NULL}

\item{summarize_obj}{sparse matrix, vector, or data.frame}

\item{summarize_by_vec}{vector with group annotation, has to have length equal to ncol(sparse_mat)/nrow(data.frame)/length(vector)}

\item{method}{method used to summarize. Only mode available for non-numeric vectors, sum/mean for sparse matrix and mode/mean/median for numeric vectors}

\item{return_matrix}{return as matrix instead of sparse matrix}

\item{order}{order the summarized vector by the groups}

\item{return_vec}{return only summarized vector instead of data.frame with summarized vector and groups}

\item{logfc_column}{sparse matrix, vector, or data.frame}

\item{pval_column}{vector with group annotation, has to have length equal to ncol(sparse_mat)/nrow(data.frame)/length(vector)}

\item{symbol_column}{provide \code{symbol_column} to add text to all points that are highlighted (when use_text=TRUE)}

\item{marker_thres}{define \code{marker_thres} to highlight all points with absolute value above}

\item{labs_range}{vector specifying range of x axis (and for fc_fc plot range of y axis)}

\item{alpha}{alpha level for points (working?)}

\item{col_vec}{define larger, smaller, intermediate and highlight colours}

\item{markers_highlight}{vector of genes from the symbol_column which should additionally be highlighted}

\item{markers_highlight_col}{named vector assigning a colour to each gene to be highlighed}

\item{only_highlight}{if TRUE, only show markers specified in \code{markers_highlight}}

\item{col_by_highlight}{if TRUE, only colour the markers specified in \code{markers_highlight}}

\item{use_text}{plot text if there are any points to highlight}

\item{add_thres_line}{if TRUE, plot dashed lines at the defined \code{marker_thres}}
}
\value{
ggplot scatterplot of log10(pval) vs FC or FC vs FC
}
\description{
plot histogram of data excluding 0 (useful for showing gene expression in sparse single cell data)
use \code{thres} to define threshold for counting number of cells above and below this mark

jj_summarize_sparse_mat calculates the sum or mean of values for each group in a sparse matrix.
jj_summarize_vector summarizes a vector based on a group vector
jj_summarize_dataframe summarizes each column in a data.frame based on a group vector

jj_volcano_plot volcano plot of pval (which is -log10 transformed) against logFC
jj_fc_fc_plot plot fold changes from two different DEG analyses against each other
}
\examples{
cd8a_count = rnbinom(1e6, size = 1, prob = 0.7)
cell_nCount_RNA = rnbinom(1e6, size = 1000, prob=0.7)
cd8a_norm_count = cd8a_count / cell_nCount_RNA
jj_plot_hist_wo0(feature_vec = cd8a_norm_count, title = 'CD8A counts', thres = 0.015)
jj_plot_hist_wo0(feature_vec = cd8a_norm_count, title = 'CD8A counts', thres = 0.015, 
                plot_numbers = c(3,2), cols = c('grey', 'black')) +
  theme(legend.position = 'none')
df = data.frame(a=seq(1,5), b=c('a','b','b','b','a'), d=c(3,0,0,1,5))
jj_summarize_vector(df$d, df$b, method='mean')
jj_summarize_dataframe(df, df$b)
marker_df = FindMarkers(pbmc_small, group.by = 'groups', ident.1 = 'g1', ident.2 = 'g2')
marker_df$symbol = rownames(marker_df)
#simple scatterplot
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', labs_range = NULL, marker_thres = Inf) + 
 theme(legend.position = 'none')
#colour highlighting above a threshold and capping very high and low values, adding lines for the threshold
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', labs_range = c(-2,2), marker_thres = 1.5, add_thres_line=T)
#adding text to highlighted markers, possibility to show unbalanced x axis
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', 
               symbol_column = 'symbol', use_text = T, labs_range = c(-2,4), marker_thres = 2)
#instead of highlighting everything above/below a threshold, specific markers can be highlighted
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', symbol_column = 'symbol',
               labs_range = c(-2.5, 2.5), marker_thres = 4,
               markers_highlight = c('FUOM','NRBP1','TAF7'), pt.size = 1.5, 
               only_highlight = F, col_by_highlight = T)
#specific highlighting and general highlighting can be combined
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', symbol_column = 'symbol',
               labs_range = c(-2.5, 2.5), marker_thres = 2,
               markers_highlight = c('FUOM','NRBP1','TAF7'), pt.size = 1.5, highlight.pt.size = 3,
               only_highlight = F, col_by_highlight = F)
#only show highlighted markers
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', symbol_column = 'symbol',
               labs_range = c(-2.5, 2.5), marker_thres = 2,
               markers_highlight = c('FUOM','NRBP1','TAF7'), pt.size = 1.5, 
               only_highlight = T, col_by_highlight = F)
#to change colour and size of highlighted markers use `highlight.pt.size` and `markers_highlight_cols`                
jj_volcano_plot(marker_df, logfc_column = 'avg_log2FC', pval_column = 'p_val', symbol_column = 'symbol',
               labs_range = c(-4, 4), marker_thres = Inf,
               markers_highlight = c('FUOM','NRBP1','TAF7'), pt.size = 1.5, highlight.pt.size = 3,
               only_highlight = F, col_by_highlight =T, 
               markers_highlight_col = structure(c('red','orange','red'), names=c('FUOM','NRBP1','TAF7'))) + theme(legend.position='none')
marker_df2 = FindMarkers(pbmc_small, group.by = 'RNA_snn_res.1', ident.1 = '0', ident.2 = '1')
marker_df2$symbol = rownames(marker_df2)
markers_joined = dplyr::inner_join(marker_df, marker_df2, by = 'symbol')
# simple logFC-logFC plot with highlighting above |avg_log2FC| = 1
jj_fc_fc_plot(plot_df = markers_joined, logfc_column1 = 'avg_log2FC.x', logfc_column2 = 'avg_log2FC.y', 
             marker_thres = 1, labs_range = NULL)    
# adding range to plot and text to markers, also adding threshold lines
jj_fc_fc_plot(markers_joined, logfc_column1 = 'avg_log2FC.x', logfc_column2 = 'avg_log2FC.y',symbol_column = 'symbol',
             labs_range = c(-4,4, -4,4), marker_thres = 2, use_text = T, add_thres_line = T) 
#specifying custom highlight
jj_fc_fc_plot(markers_joined, logfc_column1 = 'avg_log2FC.x', logfc_column2 = 'avg_log2FC.y',symbol_column = 'symbol',
             labs_range = c(-4,4, -4,4), marker_thres = 2, use_text = T,
             markers_highlight = c('RGS1','GLNY','GRN'), col_by_highlight = T) 
#changing colour and size of custom highlight
jj_fc_fc_plot(markers_joined, logfc_column1 = 'avg_log2FC.x', logfc_column2 = 'avg_log2FC.y',symbol_column = 'symbol',
             labs_range = c(-4,4, -4,4), marker_thres = Inf, use_text = T,
            markers_highlight = c('RGS1','GLNY','GRN'), col_by_highlight = T,highlight.pt.size = 3,
             markers_highlight_col = structure(rep('orange',3), names=c('RGS1','GLNY','GRN')))
}
\keyword{summarize}
