% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{timeVARS}
\alias{timeVARS}
\title{timeVARS}
\usage{
timeVARS(df, dateCol)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
timeVARS()
}
\keyword{cats}
