% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{timeVARS_DT}
\alias{timeVARS_DT}
\title{timeVARS_DT}
\usage{
timeVARS_DT(df, dateCol)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
timeVARS_DT()
}
\keyword{cats}
