% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj.R
\name{tosort}
\alias{tosort}
\title{tosort}
\usage{
tosort(datatosort, col)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
tosort()
}
\keyword{cats}
