% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zip_lists}
\alias{zip_lists}
\title{zip together two lists into one}
\usage{
zip_lists(list1, list2)
}
\arguments{
\item{list1}{list 1}

\item{list2}{list 2}

\item{listoflists}{for zip_lists_multi: a list containing multiple lists}
}
\value{
One list with alternating values from either list
}
\description{
zip two (or more) lists of equal length similar to pythons zip()
}
\examples{
zip_lists(list('A', 'B'), list(1,2))
zip_lists_multi(list(list('A', 'B'), list(1,2), list('aa','bb')))
}
