% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zip_vectors}
\alias{zip_vectors}
\title{zip vectors into one}
\usage{
zip_vectors(v1, v2)
}
\arguments{
\item{v1}{vector 1}

\item{v2}{vector 2}

\item{listofvectors}{list of vectors of equal length to zip}
}
\description{
Combine two (or more) vectors into one in alternating order
}
\examples{
zip_vectors(v1 = letters[1:3], v2 = LETTERS[1:3])
test_list = list(a=c('A','AA','AAA'), b=c('B','BB','BBB'), c=c('C', 'CC', 'CCC'))
zip_vectors_multi(test_list)
}
