% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_feature_heatmap.R
\name{jj_feature_heatmap}
\alias{jj_feature_heatmap}
\alias{jj_plot_heatmap}
\alias{jj_plot_dotplot}
\title{feature heatmap}
\usage{
jj_plot_heatmap(
  obj,
  features_use,
  group_vec,
  scale_data = TRUE,
  top_annot = NULL,
  row_annot = NULL,
  return_matrix = FALSE,
  cluster_rows = T,
  cluster_columns = T
)

jj_plot_dotplot(
  obj,
  features_use,
  group_vec,
  group_annot_df = NULL,
  scale_data = FALSE,
  column_colors = NULL,
  cluster_rows = TRUE,
  cluster_columns = TRUE
)
}
\arguments{
\item{obj}{Seurat object or matrix with features in rows and cells in columns}

\item{features_use}{Features to plot from the Seurat active assay or the matrix}

\item{group_vec}{vector with group annotation, has to have length equal to ncol(obj)}

\item{scale_data}{scale feature-wise (i.e. the columns of the heatmap)}

\item{column_colors}{}
}
\description{
feature heatmap
}
\examples{
jj_plot_heatmap(pbmc_small, features_use = c('CD8A','KLRG1','CD79A','CD79B','CD3E'), group_vec = pbmc_small$groups, scale_data=F)
jj_plot_dotplot(GetAssayData(pbmc_small), features_use = c('CD8A','KLRG1','CD79A','CD79B','CD3E'), group_vec = pbmc_small$groups)
}
