% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jj_summarize_obj.R
\name{summarise_vals}
\alias{summarise_vals}
\alias{jj_summarize_sparse_mat}
\alias{jj_summarize_vector}
\alias{jj_summarize_dataframe}
\title{summarise values as mean, mode, or sum by a group vector}
\usage{
jj_summarize_sparse_mat(
  summarize_obj,
  summarize_by_vec,
  type = "mean",
  return_matrix = TRUE
)

jj_summarize_vector(
  summarize_obj,
  summarize_by_vec,
  type = "mean",
  order = T,
  return_vec = F
)

jj_summarize_dataframe(summarize_obj, summarize_by_vec)
}
\arguments{
\item{summarize_obj}{list of ggplot objects}

\item{summarize_by_vec}{vector with group annotation, has to have length equal to ncol(sparse_mat)/nrow(data.frame)/length(vector)}

\item{return_matrix}{return as matrix instead of sparse matrix}

\item{order}{order the summarized vector by the groups}

\item{return_vec}{return only summarized vector instead of data.frame with summarized vector and groups}
}
\description{
jj_summarize_sparse_mat calculates the sum or mean of values for each group in a sparse matrix.
jj_summarize_vector summarizes a vector based on a group vector
jj_summarize_dataframe summarizes each column in a data.frame based on a group vector
}
\examples{
df = data.frame(a=seq(1,5), b=c('a','b','b','b','a'), d=c(3,0,0,1,5))
jj_summarize_vector(df$d, df$b, type='mean')
jj_summarize_dataframe(df, df$b)
}
\keyword{summarize}
